create utils package
