+++ programs/gateway/src/utils/constants.rs
@@ -1,0 +1,5 @@
+/// Prefix used for outbounds message hashes.
+pub const ZETACHAIN_PREFIX: &[u8] = b"ZETACHAIN";
+
+/// Default gas cost in lamports for SPL token operations
+pub const DEFAULT_GAS_COST: u64 = 5000;
