+++ programs/gateway/src/utils/mod.rs
@@ -1,0 +1,16 @@
+mod constants;
+mod prepare_account_metas;
+mod recover_and_verify_eth_address;
+mod validate_message;
+mod validate_message_hash;
+mod verify_and_update_nonce;
+mod verify_ata_match;
+
+// Re-export everything for easier access
+pub use constants::*;
+pub use prepare_account_metas::*;
+pub use recover_and_verify_eth_address::*;
+pub use validate_message::*;
+pub use validate_message_hash::*;
+pub use verify_and_update_nonce::*;
+pub use verify_ata_match::*;
